finally added udp trait, non function implementation right now
